requirements.txt
requests==2.31.0
python-dotenv==1.0.0
yfinance==0.2.36
pandas>=2.0.0
lxml>=4.9.0
gunicorn==21.2.0
textblob==0.17.1
aiohttp==3.9.1
beautifulsoup4>=4.12.2
vaderSentiment==3.3.2 
python-telegram-bot==13.7


numpy==1.23.5
scikit-learn==1.0.2
